102-fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,7 +9,7 @@
 int main(void)
 {
 	int a = 0;
-	int b = 1, c = 2;
+	long b = 1, c = 2;
 
 	while (a < 50)
 	{
